refactor(TabbedRules): RHICOMPL-1782 split rule selection utils
