+++ src/PresentationalComponents/TabbedRules/TabbedRules.js
@@ -7,6 +7,7 @@ import {
 import { RoutedTabs } from 'PresentationalComponents';
 import ProfileTabContent from './ProfileTabContent';
 import OsVersionText from './OsVersionText';
+import { selectedRuleRefIdsForTab } from './ruleSelection';
 
 const eventKey = ({ id, osMinorVersion }, newOsMinorVersion) => (
     `rules-${id}-${osMinorVersion || newOsMinorVersion}`
@@ -26,34 +27,6 @@ const getDefaultTab = (tabsData, defaultTab) => {
     }
 
     return eventKey(defaultTab);
-};
-
-const selectedRuleRefIdsForTab = (selectedRuleRefIds, { id }) => {
-    const tabSelection = (selectedRuleRefIds || []).find((selection) =>
-        selection.id === id
-    );
-    return tabSelection?.ruleRefIds || [];
-};
-
-export const profilesWithRulesToSelection = (profiles, prevSelection = [], options = {}) => {
-    const { only } = options;
-    const additionalSelection = profiles.map((profile) => {
-        const foundSelection = prevSelection.find(({ id }) => id === profile?.id);
-        if (!foundSelection) {
-            if (!profile.rules) {
-                console.error(`Profile ${profile.id} is missing rules for selection!`);
-            }
-
-            return {
-                id: profile.id,
-                ruleRefIds: profile.rules?.map((rule) => (rule.refId)) || []
-            };
-        } else if (only) {
-            return foundSelection;
-        }
-    }).filter((v) => !!v);
-
-    return only ? additionalSelection : [...prevSelection, ...additionalSelection];
 };
 
 const TabbedRules = ({
