+++ src/PresentationalComponents/TabbedRules/index.js
@@ -1,1 +1,2 @@
-export { default as TabbedRules, profilesWithRulesToSelection } from './TabbedRules';
+export { default as TabbedRules } from './TabbedRules';
+export { profilesWithRulesToSelection } from './ruleSelection';
